requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# LangChain and AI dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.6.1

# Vector store and embeddings
faiss-cpu==1.7.4

# Redis dependencies for persistent memory
redis==5.0.1
redis-py-cluster==2.1.3

# Data processing and utilities
pyyaml==6.0.1
pathlib2==2.3.7
typing-extensions==4.8.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2